backend/requirements.txt
# Dashboard Educativo - Backend Dependencies
# Context-Aware Implementation - Phase 1 Critical

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Cache & Background Tasks
redis==5.0.1
celery==5.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Environment
python-dotenv==1.0.0